Move ssm_chunking into ssm/ as a free function
The scan is specific to recurrent mixers, not a property every HybridStack
layer shares, so it does not belong on the block. Take layer_type_list and
layers as arguments instead.

from_model imports it locally, alongside the Symbols import already there:
ssm_inference reaches megatron.core.inference.contexts, which imports this
module, so a top-level import would close a cycle.
