+++ megatron/core/models/hybrid/hybrid_block.py
@@ -8,7 +8,7 @@
 import copy
 from contextlib import nullcontext
 from dataclasses import dataclass
-from typing import NamedTuple, Optional, Tuple, Union
+from typing import Optional, Tuple, Union
 
 import torch
 from torch import Tensor, nn
@@ -50,19 +50,6 @@ class HybridStackSubmodules:
     mlp_layer: Union[ModuleSpec, type] = IdentityOp
     moe_layer: Union[ModuleSpec, type] = IdentityOp
     mtp_block_spec: Optional[ModuleSpec] = None
-
-
-class SSMChunking(NamedTuple):
-    """Chunk-related facts shared by every SSM layer in a stack."""
-
-    chunk_size: int
-    """The mixer's configured chunk size."""
-
-    inference_chunk_size: int
-    """The chunk length the dynamic-inference prefill kernels actually run at."""
-
-    num_householder: int
-    """Householder copies for Gated Delta Product layers; 0 for other mixers."""
 
 
 class HybridStack(MegatronModule):
@@ -281,53 +268,6 @@ class HybridStack(MegatronModule):
             if layer_type == LayerSymbols.GDN:
                 return layer.self_attention.mamba_state_shapes_per_request()
         return None
-
-    def ssm_chunking(self) -> Optional["SSMChunking"]:
-        """Returns the chunking every SSM layer in this block shares, or None.
-
-        None means this block holds no recurrent layer, which happens on a
-        pipeline stage made up entirely of attention and MLP layers.
-
-        The stack is assumed homogeneous: one mixer type, one chunking. A mixed
-        stack would need a per-mixer alignment quantum and per-mixer chunk
-        descriptors, and nothing downstream models that, so it is rejected here
-        rather than silently taking the first layer's answer for every layer.
-        """
-        chunking = None
-        first_layer_idx = None
-        for layer_idx, (layer_type, layer) in enumerate(zip(self.layer_type_list, self.layers)):
-            # Mamba-family mixers (including Gated Delta Product) hang off
-            # `.mixer`; Gated Delta Net registers its recurrent mixer in the
-            # attention slot instead.
-            if layer_type == LayerSymbols.MAMBA:
-                mixer = getattr(layer, 'mixer', None)
-            elif layer_type == LayerSymbols.GDN:
-                mixer = getattr(layer, 'self_attention', None)
-            else:
-                continue
-            if mixer is None:
-                continue
-
-            layer_chunking = SSMChunking(
-                chunk_size=mixer.chunk_size,
-                # The chunk length the inference kernels actually run at, which
-                # is not always `chunk_size`: the forked Gated Delta Product
-                # prefill kernels chunk at a fixed 64. Falls back to chunk_size
-                # for any mixer predating the property.
-                inference_chunk_size=getattr(mixer, 'ssm_inference_chunk_size', mixer.chunk_size),
-                # Gated Delta Product layers register as Mamba layers but carry
-                # a Householder count, which sizes their (separate) chunk
-                # descriptors.
-                num_householder=getattr(mixer, 'num_householder', 0) or 0,
-            )
-            if chunking is None:
-                chunking, first_layer_idx = layer_chunking, layer_idx
-            else:
-                assert layer_chunking == chunking, (
-                    f"every SSM layer must share one chunking; layer {first_layer_idx} has "
-                    f"{chunking} but layer {layer_idx} has {layer_chunking}"
-                )
-        return chunking
 
     def forward(
         self,
